add contants type
